infrastructure/requirements.txt
aws-cdk-lib==2.61.1
constructs>=10.0.0,<11.0.0